[iris] nsys: make each task hook self-contained in a hooks/ package
Split iris/cluster/hooks.py into a package where each hook owns everything it
contributes to a submission, instead of reaching into setup_scripts.py and
having the CLI spell out its flags:

- hooks/nsys.py: NsysHook, its nsys_setup_script()/nsys_bin_glob() install (moved
  out of setup_scripts.py), the NSYS_* constants, and the --profile* click options
  plus build_profile_hook() — so the CLI applies one decorator and calls one
  builder, never naming an nsys flag itself.
- hooks/nsys_main.py: the run-phase launch wrapper (moved from iris/runtime/nsys.py),
  resolving its install path from the sibling hooks/nsys.py.
- hooks/multigpu.py: the MultiGpuHook spec.
- hooks/__init__.py: the TaskHook protocol and a click-free re-export of MultiGpuHook,
  so importing the package (and hence iris.cluster.types) does not pull in click.

The resource-aware builders (build_multigpu_hook, collect_hooks) stay in the client,
keeping the hooks package a leaf. The multigpu run-phase supervisor stays at
iris.runtime.multigpu — iris.runtime.jax_init consumes its rank-env contract and it
is general in-task runtime, not a hook-only entrypoint.
